refactor: move VersionEntry to types.ts, clean up CatalogFormat re-export
